Disable npm install (for now)
In order to run a sequence of shell commands, it helps to use `exec`  as opposed to `spawn`.  However `exec` has a max buffer size that npm's verbosity exceeds.

Moving to spawn has many issues involved, so for now we'll have to state that node_modules must be committed to source control.